pyproject.toml
[project]
name = "herald"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.0",
    "dotenv>=0.9.9",
    "gradio>=6.5.1",
    "langchain-text-splitters>=1.1.0",
    "openai-agents>=0.8.1",
    "pydantic>=2.12.5",
    "pymupdf4llm>=0.2.9",
    "rich>=14.3.2",
]
